Add files via upload
This catalogue expands on the Breviarium in "'What’s in a Name?’  Mapping Women’s Names from the Graffiti of Herculaneum and Pompeii" in B. Longfellow and M. Swetnam-Burland, Negotiating Silence: Material Culture and Ancient Roman Women on the Bay of Naples. This catalogue include onomastic comparanda, the full text of inscriptions, and translations of each graffito. It is sorted alphabetically.